requirements/develop.txt
-r base.txt
django-debug-toolbar==2.1
factory-boy==2.12.1
Faker==4.0.0
freezegun==0.3.13
mock==3.0.5
pytest==5.3.2
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-forked==1.1.3
pytest-sugar==0.9.2
pytest-xdist==1.31.0
